+++ src/Glutinum.Build/Main.fs
@@ -1,6 +1,7 @@
 module Build.Main
 
 open SimpleExec
+open BlackFox.CommandLine
 
 // This is a basic help message, as the CLI parser is not a "real" CLI parser
 // For now, it is enough as this is just a dev tool
@@ -12,8 +13,12 @@ Usage: dotnet run <command> [<args>]
 Available commands:
     test                            Run the main tests suite
 
-        Options for all except integration and standalone:
+        Options:
             --watch                 Watch for changes and re-run the tests
+
+    cli                             Build the CLI tool
+        Options:
+            --watch                 Watch for changes and re-build the CLI tool
 
     lint                            Run the linter on the source code
 
@@ -28,6 +33,22 @@ Available commands:
 
     printfn $"%s{helpText}"
 
+module Cli =
+
+    let handle (args: string list) =
+        let isWatch = args |> List.contains "--watch"
+
+        Command.Run(
+            "dotnet",
+            CmdLine.empty
+            |> CmdLine.appendRaw "fable"
+            |> CmdLine.appendRaw "src/Glutinum.Converter.CLI"
+            |> CmdLine.appendPrefix "--outDir" "dist"
+            |> CmdLine.appendRaw "--sourceMaps"
+            |> CmdLine.appendIf isWatch "--watch"
+            |> CmdLine.toString
+        )
+
 [<EntryPoint>]
 let main argv =
     let argv = argv |> Array.map (fun x -> x.ToLower()) |> Array.toList
@@ -39,6 +60,7 @@ let main argv =
     | "publish" :: args -> Publish.handle args
     | "lint" :: _ -> Command.Run("dotnet", "fantomas --check src")
     | "format" :: _ -> Command.Run("dotnet", "fantomas src")
+    | "cli" :: args -> Cli.handle args
     | "help" :: _
     | "--help" :: _
     | _ -> printHelp ()